+++ src/main/java/br/fiap/projeto/contexto/identificacao/domain/Cliente.java
@@ -1,0 +1,48 @@
+package br.fiap.projeto.contexto.identificacao.domain;
+
+import br.fiap.projeto.contexto.identificacao.domain.vo.Cpf;
+import br.fiap.projeto.contexto.identificacao.domain.vo.Email;
+
+public class Cliente {
+ 
+	private Long codigo;
+	 
+	private String nome;
+	 
+	private Cpf cpf;
+	 
+	private Email email;
+
+	public Long getCodigo() {
+		return codigo;
+	}
+
+	public void setCodigo(Long codigo) {
+		this.codigo = codigo;
+	}
+
+	public String getNome() {
+		return nome;
+	}
+
+	public void setNome(String nome) {
+		this.nome = nome;
+	}
+
+	public Cpf getCpf() {
+		return cpf;
+	}
+
+	public void setCpf(Cpf cpf) {
+		this.cpf = cpf;
+	}
+
+	public Email getEmail() {
+		return email;
+	}
+
+	public void setEmail(Email email) {
+		this.email = email;
+	}
+}
+
