--- src/main/java/br/fiap/projeto/contexto/identificacao/domain/port/repository/ClienteRepository.java
@@ -1,19 +1,0 @@
-package br.fiap.projeto.contexto.identificacao.domain.port.repository;
-
-import br.fiap.projeto.contexto.identificacao.domain.entity.Cliente;
-
-import java.util.List;
-import java.util.UUID;
-
-public interface ClienteRepository {
-
-    Cliente busca(UUID codigo);
-
-    List<Cliente> buscaTodos();
-
-    Cliente insere(Cliente cliente);
-
-    Cliente edita(Cliente cliente);
-
-    void remove(UUID codigo);
-}
