+++ src/main/java/br/fiap/projeto/contexto/identificacao/domain/vo/Cpf.java
@@ -1,34 +1,19 @@
 package br.fiap.projeto.contexto.identificacao.domain.vo;
 
-
-import br.fiap.projeto.exception.InvalidInputException;
-
-import java.util.regex.Pattern;
-
 public class Cpf {
-
-	public final static String CPF_INVALIDO = "Cpf inválido!";
+ 
 	private String numero;
-
-	public Cpf(String numero) {
-		this.numero = numero;
-	}
-
-	public void validar() throws InvalidInputException {
-
-		if (Pattern.matches("\\d{11}", numero)) {
-			return;
-		}
-		throw new InvalidInputException(CPF_INVALIDO);
+	 
+	public void validar() {
+	 
 	}
 
 	public String getNumero() {
 		return numero;
 	}
 
-	public static Cpf fromString(String numero) {
-
-		return new Cpf(numero);
+	public void setNumero(String numero) {
+		this.numero = numero;
 	}
 }
