--- src/main/java/br/fiap/projeto/exception/BaseException.java
@@ -1,15 +1,0 @@
-package br.fiap.projeto.exception;
-
-public class BaseException extends Exception {
-
-    private String message;
-
-    public BaseException(String message) {
-        this.message = message;
-    }
-
-    @Override
-    public String getMessage() {
-        return message;
-    }
-}
