--- src/main/java/br/fiap/projeto/exception/EntityNotFoundException.java
@@ -1,8 +1,0 @@
-package br.fiap.projeto.exception;
-
-public class EntityNotFoundException extends BaseException {
-
-    public EntityNotFoundException(String message) {
-        super(message);
-    }
-}
