--- src/main/java/br/fiap/projeto/exception/InvalidInputException.java
@@ -1,8 +1,0 @@
-package br.fiap.projeto.exception;
-
-public class InvalidInputException extends BaseException {
-
-    public InvalidInputException(String message) {
-        super(message);
-    }
-}
