+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialec
 spring.jpa.hibernate.ddl-auto=create
 spring.jpa.generate-ddl=true
 spring.jpa.hibernate.show-sql=true
-spring.datasource.url=jdbc:postgresql://localhost:5432/test_db
+spring.datasource.url=jdbc:localhost://db:5432/test_db
 spring.datasource.username=postgres
 spring.datasource.password=postgres-projeto-fiap-2023-Q1-64
