+++ src/test/java/com/fiap/projeto/AplicacaoProjetoFiapTests.java
@@ -1,0 +1,13 @@
+package com.fiap.projeto;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AplicacaoProjetoFiapTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
